fix dehacked integer parsing. original parsers allowed random text at the end of the integer so it needs to be accounted for. fix double values.

+++ Core/Dehacked/DehackedApplier.cs
@@ -465,11 +465,11 @@ public class DehackedApplier
     private static bool IsGroupValid(int? value) =>
          value == null || value.Equals(Constants.DefaultGroupNumber);
 
-    private static int GetThingSpeed(double speed, EntityDefinition definition)
+    private static int GetThingSpeed(int speed, EntityDefinition definition)
     {
         if (definition.Flags.Missile)
             return (int)GetDouble(speed);
-        return (int)speed;
+        return speed;
     }
 
     private void ApplyThingFrame(DehackedDefinition dehacked, EntityFrameTable entityFrameTable,
@@ -891,7 +891,7 @@ public class DehackedApplier
         }
     }
 
-    private static double GetDouble(double value) => value / 65536.0;
+    private static double GetDouble(int value) => value / 65536.0;
 
     private string GetSound(DehackedDefinition dehacked, int sound)
     {
